+++ lib/vcr/test_frameworks/cucumber.rb
@@ -26,18 +26,23 @@ module VCR
     # @param [Array<String>] tag_names the cucumber scenario tags
     # @param [(optional) Hash] options the cassette options. Specify :use_scenario_name => true to automatically name the cassette according to the scenario name.
     def tags(*tag_names)
-      options = tag_names.last.is_a?(::Hash) ? tag_names.pop : {}
+      original_options = tag_names.last.is_a?(::Hash) ? tag_names.pop : {}
       tag_names.each do |tag_name|
         tag_name = "@#{tag_name}" unless tag_name =~ /\A@/
-        cassette_name = "cucumber_tags/#{tag_name.gsub(/\A@/, '')}"
 
         # It would be nice to use an Around hook here, but
         # cucumber has a bug: background steps do not run
         # within an around hook.
         # https://gist.github.com/652968
         @main_object.Before(tag_name) do |scenario|
-          options = options.dup
-          cassette_name = "#{scenario.feature.name}/#{scenario.name}" if options.delete(:use_scenario_name)
+          options = original_options.dup
+
+          cassette_name = if options.delete(:use_scenario_name)
+            "#{scenario.feature.name.split("\n").first}/#{scenario.name}"
+          else
+            "cucumber_tags/#{tag_name.gsub(/\A@/, '')}"
+          end
+
           VCR.insert_cassette(cassette_name, options)
         end