+++ spec/vcr/configuration_spec.rb
@@ -26,7 +26,8 @@ describe VCR::Configuration do
       subject.default_cassette_options.should eq({
         :match_requests_on => VCR::RequestMatcherRegistry::DEFAULT_MATCHERS,
         :record            => :once,
-        :serialize_with    => :yaml
+        :serialize_with    => :yaml,
+        :persist_with      => :file_system
       })
     end
 
@@ -202,6 +203,26 @@ describe VCR::Configuration do
     end
   end
 
+  describe "#after_http_request" do
+    let(:raw_request) { VCR::Request.new }
+    let(:response)    { VCR::Response.new }
+
+    def request(type)
+      VCR::Request::Typed.new(raw_request, type)
+    end
+
+    it 'handles symbol request predicate filters properly' do
+      yielded = false
+      subject.after_http_request(:stubbed?) { |req| yielded = true }
+      subject.invoke_hook(:after_http_request, request(:stubbed), response)
+      yielded.should be_true
+
+      yielded = false
+      subject.invoke_hook(:after_http_request, request(:ignored), response)
+      yielded.should be_false
+    end
+  end
+
   describe "#around_http_request, when called on ruby 1.8" do
     it 'raises an error since fibers are not available' do
       expect {
@@ -216,6 +237,15 @@ describe VCR::Configuration do
       expect { subject.cassette_serializers[:custom] }.to raise_error(ArgumentError)
       subject.cassette_serializers[:custom] = custom_serializer
       subject.cassette_serializers[:custom].should be(custom_serializer)
+    end
+  end
+
+  describe "#cassette_persisters" do
+    let(:custom_persister) { stub }
+    it 'allows a custom persister to be registered' do
+      expect { subject.cassette_persisters[:custom] }.to raise_error(ArgumentError)
+      subject.cassette_persisters[:custom] = custom_persister
+      subject.cassette_persisters[:custom].should be(custom_persister)
     end
   end