+++ spec/vcr/middleware/faraday_spec.rb
@@ -11,14 +11,31 @@ describe VCR::Middleware::Faraday do
 
   context 'when making parallel requests' do
     include VCRStubHelpers
-    let(:parallel_manager)   { ::Faraday::Adapter::Typhoeus.setup_parallel_manager }
     let(:connection)         { ::Faraday.new { |b| b.adapter :typhoeus } }
+    let(:request_url) { "http://localhost:#{VCR::SinatraApp.port}/" }
+
+    it 'works correctly with multiple parallel requests' do
+      recorded, played_back = [1, 2].map do
+        responses = []
+
+        VCR.use_cassette("multiple_parallel") do
+          connection.in_parallel do
+            responses << connection.get(request_url)
+            responses << connection.get(request_url)
+          end
+        end
+
+        responses.map(&:body)
+      end
+
+      # there should be no blanks
+      recorded.select { |r| r.to_s == '' }.should eq([])
+      played_back.should eq(recorded)
+    end
 
     shared_examples_for "exclusive library hook" do
-      let(:request_url) { "http://localhost:#{VCR::SinatraApp.port}/" }
-
       def make_request
-        connection.in_parallel(parallel_manager) { connection.get(request_url) }
+        connection.in_parallel { connection.get(request_url) }
       end
 
       it 'makes the faraday middleware exclusively enabled for the duration of the request' do
@@ -60,7 +77,7 @@ describe VCR::Middleware::Faraday do
         undef make_request
         def make_request
           expect {
-            connection.in_parallel(parallel_manager) { connection.get(request_url) }
+            connection.in_parallel { connection.get(request_url) }
           }.to raise_error(VCR::Errors::UnhandledHTTPRequestError)
         end
       end
@@ -72,7 +89,7 @@ describe VCR::Middleware::Faraday do
       undef make_request
       def make_request(disabled = false)
         response = nil
-        connection.in_parallel(parallel_manager) do
+        connection.in_parallel do
           response = connection.get(request_url)
         end
         response