+++ CHANGELOG.md
@@ -1,6 +1,34 @@
 ## In git
 
-[Full Changelog](http://github.com/myronmarston/vcr/compare/v2.0.0...master)
+[Full Changelog](http://github.com/myronmarston/vcr/compare/v2.1.1...master)
+
+* Add new `:persist_with` cassette option. It allows you to provide a
+  customized persistence implementation so you can persist it to
+  something other than disk (i.e. a key-value store or a database).
+  Thanks to [Chris Le](https://github.com/chrisle) for the idea and
+  help with the implementation.
+* Fix `after_http_request` to handle symbol request predicate filters
+  (e.g. `:ignored?`, `:stubbed?`, `:recordable?`, `:unhandled?`, `:real?`)
+  properly. Previously using one of these would raise an ArgumentError.
+  Thanks to [playupchris](https://github.com/playupchris) for reporting
+  the bug and providing a fix.
+
+## 2.1.1 (April 24, 2012)
+
+[Full Changelog](http://github.com/myronmarston/vcr/compare/v2.1.0...v2.1.1)
+
+* Fix `:use_scenario_name` cucumber tag option so that it works properly
+  with multiple scenarios. Thanks to [Brent Snook](https://github.com/brentsnook)
+  for reporting this bug.
+* Fix `:use_scenario_name` cucumber tag option so that it only uses the
+  first line of the scenario feature name. Cucumber includes all of the
+  pre-amble text in the feature name but that can create a ridiculously
+  long cassette name. Thanks to [Brent Snook](https://github.com/brentsnook)
+  for reporting this bug.
+
+## 2.1.0 (April 19, 2012)
+
+[Full Changelog](http://github.com/myronmarston/vcr/compare/v2.0.1...v.2.1.0)
 
 * Add new `:use_scenario_name` option to the cucumber tags API. This
   allows you to use a generic tag (such as `@vcr`) and have the
@@ -13,6 +41,11 @@
   recording the cassette, in order to make it more human readable.
   Thanks to [Mislav Marohnić](https://github.com/mislav) for the
   idea and implementation.
+
+## 2.0.1 (March 30, 2012)
+
+[Full Changelog](http://github.com/myronmarston/vcr/compare/v2.0.0...v2.0.1)
+
 * Fix encoding logic to not attempt to encode the request or response
   body on deserialization if there is no encoding specified. This should
   allow cassettes recorded on 1.8 to work on 1.9. Thanks to
@@ -20,6 +53,9 @@
 * Fix Excon adapter to fix a bug with Excon 0.11 and greater. When you
   passed a block to an excon request, the response body would not be
   recorded.
+* Fix Faraday middleware so that it plays back parallel requests
+  properly. Thanks to [Dave Weiser](https://github.com/davidann) for
+  reporting this bug.
 
 ## 2.0.0 (March 2, 2012)